project/requirements.txt
Django == 3.0.5
pillow == 7.1.1
channels == 2.4.0
django-notifications-hq == 1.6.0